@@ -3,49 +3,32 @@ package edu.gatech.score.orphanconnect;
 import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Color;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
+import android.view.*;
+import android.view.animation.AnimationUtils;
+import android.widget.*;
 import androidx.annotation.NonNull;
-
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.lifecycle.ViewModelProviders;
+import androidx.recyclerview.widget.DividerItemDecoration;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import com.braintreepayments.api.dropin.DropInActivity;
 import com.braintreepayments.api.dropin.DropInRequest;
 import com.braintreepayments.api.dropin.DropInResult;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 import com.google.android.material.navigation.NavigationView;
-
-import androidx.annotation.Nullable;
-import com.nimbusds.jose.Header;
-
-import androidx.core.view.GravityCompat;
-import androidx.drawerlayout.widget.DrawerLayout;
-import androidx.appcompat.app.ActionBarDrawerToggle;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
-import androidx.recyclerview.widget.DividerItemDecoration;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 import edu.gatech.score.orphanconnect.database.OrphanListAdapter;
 import edu.gatech.score.orphanconnect.database.OrphanViewModel;
-import edu.gatech.score.orphanconnect.database.domain.Orphan;
 
-import android.view.*;
-import android.view.animation.AnimationUtils;
-import android.widget.*;
-import com.loopj.android.http.*;
-
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.sql.SQLOutput;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-
 
 public class MeetChildren extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
@@ -58,8 +41,6 @@ public class MeetChildren extends AppCompatActivity
     final String TEST_SERVER = "http://10.0.2.2:3000";
     String CLIENT_TOKEN = "sandbox_mfgm9rgj_b2rnychf57r462kj";
 
-    private OrphanViewModel orphanViewModel;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -71,16 +52,6 @@ public class MeetChildren extends AppCompatActivity
                 this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
         drawer.addDrawerListener(toggle);
         toggle.syncState();
-
-//        ImageView boy = findViewById(R.id.boy);
-//        ImageView girl = findViewById(R.id.girl);
-
-//        girl.setOnClickListener(new View.OnClickListener() {
-//            @Override
-//            public void onClick(View v) {
-//                onChildClicked(v);
-//            }
-//        });
 
         NavigationView navigationView = findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
@@ -104,7 +75,7 @@ public class MeetChildren extends AppCompatActivity
             }
         });
 
-        orphanViewModel = ViewModelProviders.of(this).get(OrphanViewModel.class);
+        OrphanViewModel orphanViewModel = ViewModelProviders.of(this).get(OrphanViewModel.class);
 
         RecyclerView recyclerView = findViewById(R.id.recyclerView);
         final OrphanListAdapter adapter = new OrphanListAdapter(this);
@@ -115,13 +86,7 @@ public class MeetChildren extends AppCompatActivity
         //Additional Code not from original guide
         recyclerView.setLayoutManager(layoutManager);
 
-        orphanViewModel.getAllOrphans().observe(this, new Observer<List<Orphan>>() {
-            @Override
-            public void onChanged(@Nullable final List<Orphan> orphans) {
-                // Update the cached copy of the words in the adapter.
-                adapter.setOrphans(orphans);
-            }
-        });
+        orphanViewModel.getAllOrphans().observe(this, adapter::setOrphans);
     }
 
     @Override
@@ -219,8 +184,6 @@ public class MeetChildren extends AppCompatActivity
                 popupWindow.dismiss();
             }
         });
-
-
 
         popupView.findViewById(R.id.donate_cancel).setOnClickListener(new View.OnClickListener() {
             @Override
